AmylaseExperiment_Simulator
This app can be used to simulate the digestion of starch with alpha-Amylase. As a result, reducing ends are liberated. These reducing ends convert DNSA->ANSA. This color change can be quantified by measuring absorbance of 540nm light.
Using this app, students can:
1) Download a template .csv file for a 96-well plate
2) Upload a 96 well plate specifying concentration of a) maltose or b) starch in each well
3) Decide whether or not to react the solution with a standardized amount of amylase
4) Specify pH either for as a) constant among all wells or b) specified for each well individually

Each time you click "run simulation, the app 'reacts' the amylase (if added) for 3 minutes with the substrate. Then, DNSA is is reacted at 90C for 10 minutes with the resulting solution. The resulting 96-well plate image (with DNSA colors) is displayed. You can then download a .csv file with the absorbance values for each well.